修改main部分逻辑
