+++ trainers/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Created by li huayong on 2020/1/6
+
+
+if __name__ == '__main__':
+    pass
